remove extension and make cargo-risczero first-class
