+++ rzup/src/lib.rs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod cargo_risczero;
 pub mod cli;
 pub mod errors;
-pub mod extension;
 pub mod repo;
 pub mod toolchain;
 pub mod utils;
